structure backend code properly
